+++ TOMB5/game/control.h
@@ -77,7 +77,7 @@ extern long tiltyoff;
 extern long OnObject;
 extern long height_type;
 
-extern long InItemControlLoop;
+extern uchar InItemControlLoop;
 extern short ItemNewRooms[256][2];
 extern short ItemNewRoomNo;
